Fixing security

+++ src/Controller/PaymentmethodController.php
@@ -9,20 +9,33 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 /**
  * @Route("/paymentmethod")
  */
 class PaymentmethodController extends AbstractController
 {
+
+    private $security;
+
+    public function __construct(Security $security)
+    {
+        $this->security = $security;
+    }
+
     /**
      * @Route("/", name="paymentmethod_index", methods={"GET"})
      */
     public function index(PaymentmethodRepository $paymentmethodRepository): Response
     {
-        return $this->render('paymentmethod/index.html.twig', [
-            'paymentmethods' => $paymentmethodRepository->findAll(),
-        ]);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            return $this->render('paymentmethod/index.html.twig', [
+                'paymentmethods' => $paymentmethodRepository->findAll(),
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 
     /**
@@ -30,22 +43,26 @@ class PaymentmethodController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $paymentmethod = new Paymentmethod();
-        $form = $this->createForm(PaymentmethodType::class, $paymentmethod);
-        $form->handleRequest($request);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            $paymentmethod = new Paymentmethod();
+            $form = $this->createForm(PaymentmethodType::class, $paymentmethod);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($paymentmethod);
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($paymentmethod);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('paymentmethod_index');
+                return $this->redirectToRoute('paymentmethod_index');
+            }
+
+            return $this->render('paymentmethod/new.html.twig', [
+                'paymentmethod' => $paymentmethod,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
         }
-
-        return $this->render('paymentmethod/new.html.twig', [
-            'paymentmethod' => $paymentmethod,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +70,13 @@ class PaymentmethodController extends AbstractController
      */
     public function show(Paymentmethod $paymentmethod): Response
     {
-        return $this->render('paymentmethod/show.html.twig', [
-            'paymentmethod' => $paymentmethod,
-        ]);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            return $this->render('paymentmethod/show.html.twig', [
+                'paymentmethod' => $paymentmethod,
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 
     /**
@@ -63,21 +84,25 @@ class PaymentmethodController extends AbstractController
      */
     public function edit(Request $request, Paymentmethod $paymentmethod): Response
     {
-        $form = $this->createForm(PaymentmethodType::class, $paymentmethod);
-        $form->handleRequest($request);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            $form = $this->createForm(PaymentmethodType::class, $paymentmethod);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('paymentmethod_index', [
-                'id' => $paymentmethod->getId(),
+                return $this->redirectToRoute('paymentmethod_index', [
+                    'id' => $paymentmethod->getId(),
+                ]);
+            }
+
+            return $this->render('paymentmethod/edit.html.twig', [
+                'paymentmethod' => $paymentmethod,
+                'form' => $form->createView(),
             ]);
+        } else {
+            return $this->redirectToRoute('default');
         }
-
-        return $this->render('paymentmethod/edit.html.twig', [
-            'paymentmethod' => $paymentmethod,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -85,12 +110,16 @@ class PaymentmethodController extends AbstractController
      */
     public function delete(Request $request, Paymentmethod $paymentmethod): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$paymentmethod->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($paymentmethod);
-            $entityManager->flush();
-        }
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            if ($this->isCsrfTokenValid('delete'.$paymentmethod->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($paymentmethod);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('paymentmethod_index');
+            return $this->redirectToRoute('paymentmethod_index');
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 }
